refactor: image url type 수정
