+++ @types/modal.d.ts
@@ -1,19 +1,19 @@
 import { Dispatch, SetStateAction } from 'react';
-import { SimpleClubData } from './club';
+import { ClubSearchData } from './club';
 
 export interface ModalProps {
   setIsOpen: Dispatch<SetStateAction<boolean>>;
   inputRef: RefObject<HTMLDivElement>;
   type: string;
   label: string;
-  searchData: SimpleClubData[];
-  setSelectedId: Dispatch<SetStateAction<SimpleClubData | null>>;
+  searchData: ClubSearchData[];
+  setSelectedId: Dispatch<SetStateAction<ClubSearchData | null>>;
 }
 
 export interface ModalContainerProps {
   handleCloseModal: () => void;
   type: string;
   label: string;
-  searchData: SimpleClubData[];
-  setSelectedId: Dispatch<SetStateAction<SimpleClubData | null>>;
+  searchData: ClubSearchData[];
+  setSelectedId: Dispatch<SetStateAction<ClubSearchData | null>>;
 }
